chore: 审查第三批 - #74 杂项清理 (#77)
* chore(model): review batch 3 - misc hardening/cleanup (#74)

Closes the remaining offline misc items from the full-codebase review:

- paged_kv: decode_step borrows the slot's page table instead of cloning it
  every step (O(seq) -> O(1) per step); token index is bounds-checked against
  vocab before the embedding lookup; PageAllocator::free guards double-free /
  out-of-range ids; GpuBlockTableBuilder asserts a fresh tail page is not
  already cached before inserting; rename paged_kv::BlockTable -> PagedTable
  to disambiguate from kv_block_pool::BlockTable.
- prefix_kv: document exactly why the non-Fresh-beyond-prefix assert is
  unreachable today (no release() path; eviction always fills the pool before a
  re-request can dedup a hole survivor) and add a heavy-eviction regression
  test proving recompute stays bit-exact.
- state_reuse: find_reusable checks the length-vs-best gate before the slice
  compare (skip dominated anchors).
- prefix_cache: hex_to_bytes asserts canonical even-length input; CacheKey
  page_offset is usize (no i32 truncation for >2^31-token contexts).
- kernels: f16_gqa geometry guard also requires head_dim % 8 == 0 (uint4
  vector reads need 16-byte aligned K rows); SAFETY comment now states the
  real confinement invariant (owners confine to one engine thread; compile
  cache is Arc-shared).
- server: remove dead duplicate storage-FP8 printlns in the Q4 branches;
  invalid MACH_DTYPE exits with a message instead of panicking; spec-mode
  error wording no longer claims a second f32 model (draft dtype follows
  MACH_DTYPE).
- sampling: document the MAX_TOPK clamp as the defensive fallback (the OpenAI
  layer already 400s top_logprobs > 20).

Gates: cargo test -p mach-model (159 lib + integration all green),
cargo test -p mach-server --features hip (12), offline 42-kernel hiprtc gate,
cargo check/clippy --workspace --features hip --all-targets clean.

* chore: review follow-ups - move vocab check before page alloc, sync renamed builder in docs, precise test comment

+++ crates/mach-model/src/reuse_planner.rs
@@ -133,7 +133,7 @@ impl ReusePlanner {
                 namespace_id: self.namespace_id,
                 group_id: self.group_id,
                 content_hash: h.clone(),
-                page_offset: (i * self.tokens_per_page) as i32,
+                page_offset: i * self.tokens_per_page,
             })
             .collect();
 
@@ -250,7 +250,7 @@ mod tests {
                 namespace_id: 7,
                 group_id: 0,
                 content_hash: h.clone(),
-                page_offset: (i * page) as i32,
+                page_offset: i * page,
             })
             .collect()
     }
